Change id film by checksum (type long) Regenerate new myBatis files. Add new IFileService and new methods.

+++ src/test/java/net/juancarlosfernandez/myfilms/poc/mybatis/Main.java
@@ -32,7 +32,7 @@ public class Main {
 				sqlSession = sqlMapper.openSession(DBUtils.getConnection(DB_DRIVER, DB_URL));
 				FilmMapper filmMapper = sqlSession.getMapper(FilmMapper.class);
 				
-				Film film = filmMapper.selectByPrimaryKey(1);
+				Film film = filmMapper.selectByPrimaryKey((long)1);
 				
 				System.out.println("Id_Film " + film.getIdFilm());
